Usage: Show optional/mandatory arguments in usage
For options with an assignment using ":" or "=", include a single
letter argument in the usage information.
  Also introduce a function "_option_length" to compute the spacing
needed for a given option (spec).  It computes the space needed more
precisely than before (which was more of a heurestic).  Particularly,
for "single long" options, there is no longer added space for a short
option.

Caveat: It ignores the "{N, M}" part causing the usage to suggest
such options only take one argument.