Adapt for changes in Getopt::Long warning output
In Getopt-Long-2.55 a long-standing bug was fixed, but this had a side
effect of emitting a warning where none was previously emitted.
Consequently, tests in t/descriptive.t which depended on a count of
warnings being emitted need to be adjusted.

See discussion in https://github.com/Perl/perl5/issues/21640.  Also note
that, per https://metacpan.org/dist/Getopt-Long/changes, this new warning will one day become a fatal error.  So this patch is mainly intended to stanch the bleeding on CPANtesters.